src/requirements.txt
python-gitlab
python-dotenv
azure-identity
openai
pyyaml
mcp_client